requirements.txt
boto3==1.6.3
botocore==1.9.3
docutils==0.14
jmespath==0.9.3
kafka-python==1.4.1
mysql-replication==0.18
PyMySQL==0.8.0
python-dateutil==2.6.1
s3transfer==0.1.13
six==1.11.0
